Updated ModerateDetectionsByIdAsync to account for multiple Ids, which is needed for bulk update. Updated unit tests for said method and supporting methods.

+++ ModeratorFrontEnd/OrcaHello/OrcaHello.Web.Api/Services/Orchestrations/Detections/DetectionOrchestrationService.cs
@@ -89,39 +89,58 @@
             return AsDetection(maybeMetadata);
         });
 
-        public ValueTask<Detection> ModerateDetectionByIdAsync(string id, ModerateDetectionRequest request) =>
+        public ValueTask<ModerateDetectionsResponse> ModerateDetectionsByIdAsync(ModerateDetectionsRequest request) =>
         TryCatch(async () =>
         {
-            Validate(id, nameof(id));
             ValidateModerateRequestOnUpdate(request);
 
-            // Get the current record
-            Metadata existingRecord = await _metadataService.RetrieveMetadataByIdAsync(id);
-            ValidateStorageMetadata(existingRecord, id);
-
-            var existingState = existingRecord.State;
-
-            // Make updates so they can be added as a new record
-            Metadata newRecord = existingRecord;
-            newRecord.State = request.State;
-            newRecord.Moderator = request.Moderator;
-            newRecord.DateModerated = request.DateModerated.ToString();
-            newRecord.Comments = request.Comments;
-            newRecord.Tags = request.Tags;
-
-            bool existingRecordDeleted = await _metadataService.RemoveMetadataByIdAndStateAsync(id, existingState);
-            ValidateDeleted(existingRecordDeleted, id);
-
-            bool newRecordCreated = await _metadataService.AddMetadataAsync(newRecord);
-
-            if(!newRecordCreated)
+            ModerateDetectionsResponse response = new()
             {
-                bool existingRecordRecreated = await _metadataService.AddMetadataAsync(existingRecord);
-                ValidateInserted(existingRecordRecreated, id);
+                IdsToUpdate = request.Ids,
+            };
+
+            foreach (var id in request.Ids)
+            {
+                // Get the current record
+                Metadata existingRecord = await _metadataService.RetrieveMetadataByIdAsync(id);
+
+                if (existingRecord is null)
+                    response.IdsNotFound.Add(id);
+                else
+                {
+                    var existingState = existingRecord.State;
+
+                    // Make updates so they can be added as a new record
+                    Metadata newRecord = existingRecord;
+                    newRecord.State = request.State;
+                    newRecord.Moderator = request.Moderator;
+                    newRecord.DateModerated = request.DateModerated.ToString();
+                    newRecord.Comments = request.Comments;
+                    newRecord.Tags = request.Tags;
+
+                    bool existingRecordDeleted = await _metadataService.RemoveMetadataByIdAndStateAsync(id, existingState);
+
+                    if (!existingRecordDeleted)
+                        response.IdsUnsuccessfullyUpdated.Add(id);
+                    else
+                    {
+                        bool newRecordCreated = await _metadataService.AddMetadataAsync(newRecord);
+
+                        if (!newRecordCreated)
+                        {
+                            response.IdsUnsuccessfullyUpdated.Add(id);
+                            bool existingRecordRecreated = await _metadataService.AddMetadataAsync(existingRecord);
+                        }
+                        else
+                        {
+                            response.IdsSuccessfullyUpdated.Add(id);
+                        }
+                    }
+                }
             }
 
-            return AsDetection(newRecord);
-        });
+            return response;
+         });
 
         public ValueTask<DetectionListForInterestLabelResponse> RetrieveDetectionsForGivenInterestLabelAsync(string interestLabel) =>
          TryCatch(async () =>
